univariate/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
matplotlib>=3.4.0

# GluonTS and time series
gluonts>=0.13.0
statsmodels>=0.12.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0

# Optional for notebook support
jupyter>=1.0.0
ipykernel>=6.0.0